Final commit.
Parameters working for setpoints 25cm-50cm.
README.md updated.
Video demonstrations for each setpoint added.
Screenshots from demonstrations imported to "MatlabCode" folder.

+++ Project/Project/src/GlobalVariables.h
@@ -2,7 +2,7 @@
  * GlobalVariables.h
  *
  * Created: 11/29/2016 13:29:01
- *  Author: Nadia, Elvin
+ * Author: Nadia Manoh and Elvin Lapovski
  */ 
 
 
@@ -10,7 +10,7 @@
 #define GLOBALVARIABLES_H_
 
 /* Hold variables that is sent between Arduino and MATLAB as well as between PIDControl.c and MatlabPlot.c */
-extern uint16_t setpoint;
+extern uint8_t setpoint;
 extern double dT;
 extern double Kp;
 extern double Ti;
